LibGfx: Make FastBoxBlurFilter::apply_single_pass's argument unsigned
It isn't sensible to have a negative radius for blurring, so an unsigned
value is the right thing to do here.
Now we have to cast the radius to int a few times when actually doing
the calculations but I'm sure that can be done more intelligently, but
that optimization is a thing for the future.

It looked very goofy for the two different ways of invoking the Filter
to have differently signed arguments.

+++ Userland/Libraries/LibGfx/Filters/FastBoxBlurFilter.h
@@ -14,7 +14,7 @@ class FastBoxBlurFilter {
 public:
     FastBoxBlurFilter(Bitmap&);
 
-    void apply_single_pass(int radius);
+    void apply_single_pass(size_t radius);
     void apply_three_passes(size_t radius);
 
 private:
